pyproject.toml
[project]
name = "app-agentic-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "google-genai>=1.16.1",
    "letta-client>=0.1.198",
    "loguru>=0.7.3",
    "redis>=6.1.0",
    "taskipy>=1.14.1",
    "uvicorn>=0.34.2",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.12.0",
    "gunicorn>=23.0.0",
    "pandas>=2.2.3",
    "mlflow>=2.22.0",
    "google-cloud>=0.34.0",
    "google-cloud-storage>=3.1.0",
    "arize-phoenix-otel>=0.9.2",
    "arize-phoenix-evals>=0.20.6",
    "arize-phoenix==8.32.1",
    "nest-asyncio>=1.6.0",
    "google-cloud-bigquery>=3.34.0",
    "google-cloud-bigquery-storage>=2.32.0",
    "google-auth>=2.40.1",
    "python-dotenv>=1.1.0",
    "typesense>=1.1.1",
    "google-generativeai>=0.8.5",
    "openai>=1.91.0",
    "httpx>=0.25.0",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.5.0",
    "langchain-postgres>=0.0.15",
    "psycopg-binary>=3.2.9",
    "langchain-mcp-adapters>=0.1.9",
    "pgvector>=0.3.6",
]

[tool.taskipy.tasks]
ngrok = "ngrok http --url=ideal-burro-barely.ngrok-free.app 8089"
serve = "uvicorn src.main:app --reload --host 0.0.0.0 --port 8089 --log-level info"
db-init = "python -c \"from src.db import Base, engine; Base.metadata.create_all(bind=engine)\""
db-migrate = "alembic revision --autogenerate -m"
db-upgrade = "alembic upgrade head"
db-downgrade = "alembic downgrade"



[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
